internal/k8s refactored
