+++ cluster-operator/controllers/vmc/sync_manifest_secret.go
@@ -11,7 +11,6 @@ import (
 	constants2 "github.com/verrazzano/verrazzano/pkg/mcconstants"
 	clusterapi "github.com/verrazzano/verrazzano/platform-operator/apis/clusters/v1alpha1"
 	"github.com/verrazzano/verrazzano/platform-operator/constants"
-	"github.com/verrazzano/verrazzano/platform-operator/controllers/clusters"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	corev1 "k8s.io/api/core/v1"
@@ -60,7 +59,7 @@ func (r *VerrazzanoManagedClusterReconciler) syncManifestSecret(ctx context.Cont
 	// attempt to fetch the registration manifest YAML
 	vzVMCWaitingForClusterID := false
 
-	if vmc.Labels != nil && vmc.Labels[clusters.createdByLabel] == clusters.createdByVerrazzano && len(vmc.Status.RancherRegistration.ClusterID) == 0 {
+	if vmc.Labels != nil && vmc.Labels[createdByLabel] == createdByVerrazzano && len(vmc.Status.RancherRegistration.ClusterID) == 0 {
 		r.log.Progressf("Waiting for Verrazzano-created VMC named %s to have a cluster id in the status before attempting to fetch Rancher registration manifest", vmc.Name)
 		vzVMCWaitingForClusterID = true
 	} else {
